# contracts/crm-pipeline/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CRM Pipeline Service",
    "description": "CRM Pipeline: leads, opportunities, pipeline stages, activities, and sales handoff.",
    "license": { "name": "" },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/crm-pipeline/opportunities": {
      "get": {
        "tags": ["Opportunities"],
        "summary": "List opportunities",
        "operationId": "list_opportunities",
        "parameters": [
          { "name": "stage_code", "in": "query", "schema": { "type": "string" } },
          { "name": "assigned_user_id", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": { "description": "Paginated list of opportunities", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Opportunity" } } } } }
        },
        "security": [{ "bearer": [] }]
      },
      "post": {
        "tags": ["Opportunities"],
        "summary": "Create an opportunity",
        "operationId": "create_opportunity",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateOpportunityRequest" } } } },
        "responses": {
          "201": { "description": "Opportunity created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Opportunity" } } } }
        },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/opportunities/{id}": {
      "get": {
        "tags": ["Opportunities"],
        "summary": "Get an opportunity",
        "operationId": "get_opportunity",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Opportunity detail", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Opportunity" } } } }
        },
        "security": [{ "bearer": [] }]
      },
      "put": {
        "tags": ["Opportunities"],
        "summary": "Update an opportunity",
        "operationId": "update_opportunity",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateOpportunityRequest" } } } },
        "responses": { "200": { "description": "Opportunity updated" } },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/opportunities/{id}/advance-stage": {
      "post": {
        "tags": ["Opportunities"],
        "summary": "Advance opportunity to next stage",
        "operationId": "advance_stage",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdvanceStageRequest" } } } },
        "responses": {
          "200": { "description": "Stage advanced" },
          "422": { "description": "Target stage is terminal or does not exist" }
        },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/opportunities/{id}/close-won": {
      "post": {
        "tags": ["Opportunities"],
        "summary": "Close opportunity as won",
        "operationId": "close_won",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CloseWonRequest" } } } },
        "responses": { "200": { "description": "Opportunity closed won" } },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/opportunities/{id}/close-lost": {
      "post": {
        "tags": ["Opportunities"],
        "summary": "Close opportunity as lost",
        "operationId": "close_lost",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CloseLostRequest" } } } },
        "responses": { "200": { "description": "Opportunity closed lost" } },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/opportunities/{id}/stage-history": {
      "get": {
        "tags": ["Opportunities"],
        "summary": "Get stage history for an opportunity",
        "operationId": "stage_history",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Stage history", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/OpportunityStageHistory" } } } } }
        },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/pipeline/summary": {
      "get": {
        "tags": ["Opportunities"],
        "summary": "Pipeline summary by stage",
        "operationId": "pipeline_summary",
        "responses": {
          "200": { "description": "Pipeline summary", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/PipelineSummaryItem" } } } } }
        },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/stages": {
      "get": {
        "tags": ["PipelineStages"],
        "summary": "List pipeline stages",
        "operationId": "list_stages",
        "responses": {
          "200": { "description": "Pipeline stages", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/PipelineStage" } } } } }
        },
        "security": [{ "bearer": [] }]
      },
      "post": {
        "tags": ["PipelineStages"],
        "summary": "Create a pipeline stage",
        "operationId": "create_stage",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateStageRequest" } } } },
        "responses": {
          "201": { "description": "Stage created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PipelineStage" } } } }
        },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/stages/{code}": {
      "put": {
        "tags": ["PipelineStages"],
        "summary": "Update a pipeline stage",
        "operationId": "update_stage",
        "parameters": [{ "name": "code", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateStageRequest" } } } },
        "responses": { "200": { "description": "Stage updated" } },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/stages/{code}/deactivate": {
      "post": {
        "tags": ["PipelineStages"],
        "summary": "Deactivate a pipeline stage",
        "operationId": "deactivate_stage",
        "parameters": [{ "name": "code", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Stage deactivated" } },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/stages/reorder": {
      "post": {
        "tags": ["PipelineStages"],
        "summary": "Reorder pipeline stages",
        "operationId": "reorder_stages",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReorderStagesRequest" } } } },
        "responses": { "200": { "description": "Stages reordered" } },
        "security": [{ "bearer": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Opportunity": {
        "type": "object",
        "required": ["id", "tenant_id", "opp_number", "title", "stage_code", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "tenant_id": { "type": "string" },
          "opp_number": { "type": "string" },
          "title": { "type": "string" },
          "stage_code": { "type": "string" },
          "opp_type": { "type": ["string", "null"] },
          "priority": { "type": ["string", "null"] },
          "party_id": { "type": ["string", "null"], "format": "uuid" },
          "primary_contact_id": { "type": ["string", "null"], "format": "uuid" },
          "assigned_user_id": { "type": ["string", "null"], "format": "uuid" },
          "estimated_value_cents": { "type": ["integer", "null"] },
          "currency_code": { "type": ["string", "null"] },
          "expected_close_date": { "type": ["string", "null"], "format": "date" },
          "close_reason": { "type": ["string", "null"] },
          "source_lead_id": { "type": ["string", "null"], "format": "uuid" },
          "sales_order_id": { "type": ["string", "null"], "format": "uuid" },
          "notes": { "type": ["string", "null"] },
          "closed_at": { "type": ["string", "null"], "format": "date-time" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "OpportunityStageHistory": {
        "type": "object",
        "required": ["id", "opportunity_id", "stage_code", "entered_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "opportunity_id": { "type": "string", "format": "uuid" },
          "stage_code": { "type": "string" },
          "entered_at": { "type": "string", "format": "date-time" },
          "entered_by_user_id": { "type": ["string", "null"], "format": "uuid" }
        }
      },
      "PipelineSummaryItem": {
        "type": "object",
        "required": ["stage_code", "display_label", "order_rank", "count", "total_value_cents"],
        "properties": {
          "stage_code": { "type": "string" },
          "display_label": { "type": "string" },
          "order_rank": { "type": "integer" },
          "count": { "type": "integer" },
          "total_value_cents": { "type": "integer" }
        }
      },
      "CreateOpportunityRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": { "type": "string" },
          "stage_code": { "type": ["string", "null"] },
          "opp_type": { "type": ["string", "null"] },
          "priority": { "type": ["string", "null"] },
          "party_id": { "type": ["string", "null"], "format": "uuid" },
          "primary_contact_id": { "type": ["string", "null"], "format": "uuid" },
          "assigned_user_id": { "type": ["string", "null"], "format": "uuid" },
          "estimated_value_cents": { "type": ["integer", "null"] },
          "currency_code": { "type": ["string", "null"] },
          "expected_close_date": { "type": ["string", "null"], "format": "date" },
          "source_lead_id": { "type": ["string", "null"], "format": "uuid" },
          "notes": { "type": ["string", "null"] }
        }
      },
      "UpdateOpportunityRequest": {
        "type": "object",
        "properties": {
          "title": { "type": ["string", "null"] },
          "opp_type": { "type": ["string", "null"] },
          "priority": { "type": ["string", "null"] },
          "party_id": { "type": ["string", "null"], "format": "uuid" },
          "primary_contact_id": { "type": ["string", "null"], "format": "uuid" },
          "assigned_user_id": { "type": ["string", "null"], "format": "uuid" },
          "estimated_value_cents": { "type": ["integer", "null"] },
          "currency_code": { "type": ["string", "null"] },
          "expected_close_date": { "type": ["string", "null"], "format": "date" },
          "notes": { "type": ["string", "null"] }
        }
      },
      "AdvanceStageRequest": {
        "type": "object",
        "required": ["stage_code"],
        "properties": {
          "stage_code": { "type": "string" }
        }
      },
      "CloseWonRequest": {
        "type": "object",
        "properties": {
          "notes": { "type": ["string", "null"] }
        }
      },
      "CloseLostRequest": {
        "type": "object",
        "required": ["close_reason"],
        "properties": {
          "close_reason": { "type": "string" },
          "notes": { "type": ["string", "null"] }
        }
      },
      "PipelineStage": {
        "type": "object",
        "required": ["id", "tenant_id", "stage_code", "display_label", "order_rank", "is_terminal", "is_win", "active", "created_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "tenant_id": { "type": "string" },
          "stage_code": { "type": "string" },
          "display_label": { "type": "string" },
          "order_rank": { "type": "integer" },
          "is_terminal": { "type": "boolean" },
          "is_win": { "type": "boolean" },
          "probability_default_pct": { "type": ["integer", "null"] },
          "active": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "CreateStageRequest": {
        "type": "object",
        "required": ["stage_code", "display_label", "order_rank", "is_terminal", "is_win"],
        "properties": {
          "stage_code": { "type": "string" },
          "display_label": { "type": "string" },
          "order_rank": { "type": "integer" },
          "is_terminal": { "type": "boolean" },
          "is_win": { "type": "boolean" },
          "probability_default_pct": { "type": ["integer", "null"] }
        }
      },
      "UpdateStageRequest": {
        "type": "object",
        "properties": {
          "display_label": { "type": ["string", "null"] },
          "order_rank": { "type": ["integer", "null"] },
          "probability_default_pct": { "type": ["integer", "null"] }
        }
      },
      "ReorderStagesRequest": {
        "type": "object",
        "required": ["stages"],
        "properties": {
          "stages": { "type": "array", "items": { "$ref": "#/components/schemas/StageReorderItem" } }
        }
      },
      "StageReorderItem": {
        "type": "object",
        "required": ["stage_code", "order_rank"],
        "properties": {
          "stage_code": { "type": "string" },
          "order_rank": { "type": "integer" }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}